refactor(api): move user-orga-settings-repository
to team bounded context
