+++ api/src/team/infrastructure/repositories/user-orga-settings-repository.js
@@ -1,8 +1,8 @@
-import { knex } from '../../../../../db/knex-database-connection.js';
-import { UserOrgaSettingsCreationError } from '../../../../../lib/domain/errors.js';
-import { Organization, User } from '../../../../../lib/domain/models/index.js';
-import { UserOrgaSettings } from '../../../../../lib/domain/models/UserOrgaSettings.js';
-import * as knexUtils from '../../../infrastructure/utils/knex-utils.js';
+import { knex } from '../../../../db/knex-database-connection.js';
+import { UserOrgaSettingsCreationError } from '../../../../lib/domain/errors.js';
+import { Organization, User } from '../../../../lib/domain/models/index.js';
+import { UserOrgaSettings } from '../../../../lib/domain/models/UserOrgaSettings.js';
+import * as knexUtils from '../../../shared/infrastructure/utils/knex-utils.js';
 
 const findOneByUserId = async function (userId) {
   const userOrgaSettings = await knex('user-orga-settings').where({ userId }).first();
